+++ src/iTin.Core.Hardware/Specification/DMI.cs
@@ -30,15 +30,10 @@ namespace iTin.Core.Hardware.Specification
         #endregion
 
         #region public static readonly properties
-        /// <summary>
-        /// Gets a unique instance of this class.
-        /// </summary>
-        public static readonly DMI Instance = new DMI();
-        #endregion
 
-        #region public static properties
+        #region [public] {static} (string) AccessType: Gets a string that represents the type of access
         /// <summary>
-        /// Gets a string that represents the type of access.
+        /// Gets a <see cref="T:System.String"/> that represents the type of access.
         /// </summary>
         /// <value>
         /// A string that represents the type of access.
@@ -47,20 +42,36 @@ namespace iTin.Core.Hardware.Specification
         /// This method always returns the <b>System BIOS</b> string.
         /// </remarks>       
         public static string AccessType => "System BIOS";
+        #endregion
 
+        #region [public] {static} (string) Identificationmethod: Gets a string that represents access mode
         /// <summary>
-        /// Obtiene una cadena que representa modo del acceso.
+        /// Gets a <see cref="T:System.String"/> that represents access mode.
         /// </summary>
         /// <value>
-        /// Una cadena que representa el modo del acceso.
+        /// A <see cref="T:System.String"/> that represents the mode of access.
         /// </value>
         /// <remarks>
-        ///Este método devuelve siempre la cadena <b>DMI</b>.
+        /// This method always returns the <b>&lt;DMI&gt;</b> string.
         /// </remarks>       
         public static string Identificationmethod => "<DMI>";
         #endregion
 
-        #region public properties
+        #region [public] {static} (DMI) Instance: Gets a unique instance of this class
+        /// <summary>
+        /// Gets a unique instance of this class.
+        /// </summary>
+        /// <value>
+        /// A unique <see cref="DMI"/> reference that contains <b>DMI</b> information.
+        /// </value>
+        public static readonly DMI Instance = new DMI();
+        #endregion
+
+        #endregion
+
+        #region public readonly properties
+
+        #region [public] (DmiStructureCollection) Structures: Gets the collection of available structure
         /// <summary>
         /// Gets the collection of available structures.
         /// </summary>
@@ -69,6 +80,8 @@ namespace iTin.Core.Hardware.Specification
         /// If there is no object <see cref="DmiStructure" />, <b>null</b> is returned.
         /// </value>
         public DmiStructureCollection Structures => dmiStructureCollection ?? (dmiStructureCollection = new DmiStructureCollection());
+        #endregion
+
         #endregion
 
         #region public overrides methods